// scripts/types.ts
export type RFOXMetadata = {
  /** The current epoch number */
  epoch: number
  /** The start timestamp for the current epoch */
  epochStartTimestamp: number
  /** The end timestamp for the current epoch */
  epochEndTimestamp: number
  /** The current percentage of revenue (RUNE) earned by the treasury to be distributed as rewards */
  distributionRate: number
  /** The current percentage of revenue (RUNE) earned by the treasury to be used to buy FOX from the open market and subsequently burned */
  burnRate: number
  /** The treasury address on THORChain used to determine revenue earned by the DAO for RFOX reward distributions and total burn */
  treasuryAddress: string
  /** A record of epoch number to their corresponding IPFS hashes */
  ipfsHashByEpoch: Record<number, string>
}
// ...
export type RewardDistribution = {
  /** The amount (RUNE) distributed to the reward address */
  amount: string
  /** The RFOX staking reward units used to calculate the reward distribution */
  rewardUnits: string
  /** The transaction ID (THORChain) for the reward distribution */
  txId: string
  /** The address used for the reward distribution */
  rewardAddress: string
}
// ...
export type Epoch = {
  /** The epoch number for this epoch */
  number: number
  /** The start timestamp for this epoch */
  startTimestamp: number
  /** The end timestamp for this epoch */
  endTimestamp: number
  /** The start block for this epoch */
  startBlock: number
  /** The end block for this epoch */
  endBlock: number
  /** The total revenue (RUNE) earned by the treasury for this epoch */
  totalRevenue: string
  /** The total RFOX staking reward units for this epoch */
  totalRewardUnits: string
  /** The percentage of revenue (RUNE) accumulated by the treasury to be distributed as rewards for this epoch */
  distributionRate: number
  /** The percentage of revenue (RUNE) accumulated by the treasury to be used to buy FOX from the open market and subsequently burned for this epoch */
  burnRate: number
  /** A record of staking address to reward distribution for this epoch */
  distributionsByStakingAddress: Record<string, RewardDistribution>
}